requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2024.12.14
dnspython==2.7.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
yarl==1.18.3
selenium==4.28.1
webdriver_manager==4.0.2
requests==2.32.3
psycopg2-binary==2.9.10
asyncpg==0.30.0


pytest~=8.3.4
pytest-asyncio==0.25.3